feat!: line number segment and config options
Add a global VimL function `ScLn()` which can be used in a `%{}` eval
block and configured through the new `thousands`, `relculright` config
options. It can also be customized further with a custom lua function
through the `lnumfunc` and `reeval` config options. The `statuscolumn`
config option can be used to let the plugin set your `'statuscolumn'`
string, following the segment order set in the `order` config option.